# feat: implement Read Buffer run-length encoding for scalars