--- a/read_buffer/src/column/encoding/scalar.rs
+++ b/read_buffer/src/column/encoding/scalar.rs
@@ -1,5 +1,7 @@
 pub mod fixed;
 pub mod fixed_null;
+pub mod rle;
 
 pub use fixed::Fixed;
 pub use fixed_null::FixedNull;
+pub use rle::RLE;