This PR introduces a new encoding to the Read Buffer: run-length encoding of scalar values.

## Rationale

As the primary in-memory execution engine within IOx the Read Buffer needs to aggresively compress data to maximise the total number of rows that can be held. However, this must be done whilst balancing read performance; compressions schemes that result in significant expense to access values harm query execution performance.

This PR adds a new RLE encoding scheme for compressing columns of scalar values (integers and floats). It allows us to significantly reduce the footprint of columns of these values for time-series data where it is often the case that there are many runs of repeated values.

Further, the encoding scheme compliments the existing design of the Read Buffer, which uses compressed bitsets to hold intermediate row offsets when applying predicates across columns within a table. Within RLE encoded columns, all row IDs for values matching predicates can be efficiently added to bitset results using range-based operations, keeping intermediate state compressed for longer.

## Background

Up to now a lot of the focus within the Read Buffer has been on compressing string (tag) columns for the following reasons:

 - each value is of variable size;
 - often tags have relatively low cardinality compared to the number of rows in a column;
 - compression allows for execution directly on the compressed representations; and 
 - compressed representations are integers which opens up more vectorised CPU instructions (SIMD).

Since IOx has been deployed within one of our internal Cloud 2 clusters we have been able to feed it a significant volume of meaningful time-series data. An analysis of that has shown that there is a significant amount of scalar data that is being stored with little compression. Further, a lot of this data is very sparse in nature, and due to the existing scalar encodings is stored uncompressed with `NULL` values taking up as much space as non-null values.

Therfore, by run-length encoding this data where it makes sense to we hope to see significant reductions without making query performance significantly worse. Indeed, in many cases predicate evaluation will be _improved_ because we can evaluate predicates directly on the compressed representation.

## Implementation Notes

 - The RLE implementation is generic over all of the logical scalar types we support (`i64`, `f64` etc), but it's only wired up to support float columns right now. I will extend support in following PRs.
 - The implementation is relatively straightforward, e.g., I have used `(u32, Option<T>)` for storing run-lengths. Naturally that can be improved upon, but this PR is following the 80/20 rule and trying to get the majority of the value added more quickly.
 - I have not wired up aggregate functions within this encoding. I will do that for completeness, but since DataFusion can't yet push aggregates down to the Read Buffer it's not vital yet.
 - I have added some heuristics to decide when to RLE a float column. We can tweak and improve these over time, but for now - if applying RLE to a float column reduces the number of rows by 30% or more then that encoding gets chosen.
 - Recently we have been thinking about how we want ordering/predicate evaluation/aggregation to happen with respect to certain floating point values (`NaN`, `infinity`, `-infinity`). This time around I have written tests that will assert that the Read Buffer behaves in the same way as PostgreSQL in terms of predicate evaluation. However, I have left the assertions in the test commented out because we need customer comparators and that isn't necessary for what this PR is trying to achieve.

## Empirical Testing

Internally we have a known "problematic" table in IOx, which presents a pathological case for sparse scalar encodings with no compression, which is that we have to spend 8 bytes on every NULL value in a column. Some columns within this table have `97-99%` sparsity (NULL values). Therefore compressiong those NULL values down with RLE results in a significantly smaller footprint.

I have tested this PR on an extract of that dataset:

### Compression

|  | Before | RLE PR |
| --- | --- | -- |
| Float columns | 47 | ..
| Float values | 7,074,111 | ..
| Sparsity (NULLS) | 6,923,598 (97%) | ..
| Size (Bytes) | 58,139,752 | **6,331,768**

For this dataset the RLE scheme compresses the columnar data by **89%** from `~58MB~ to `~6.3MB`.

### Predicate evaluation

I also tested how quickly a column encoded with this scalar run-lengtrh encoding was able to find all rows that satisfied a predicate compared to the existing scheme where float values were effectively held in a `Vec<T>`.

I tested the equivalent of following query against the above dataset:

```
SELECT "container_file_descriptors" FROM "prometheus" WHERE "container_file_descriptors" > 300;
```

There are `150,513` values in the "container_file_descriptors" column in the range `[0, 701]`, however `~99%` of them are `NULL` so the query is very selective. In total `27` rows were returned.

|  | Before | RLE PR |
| --- | --- | -- |
| Execution Time | `630μs` | `70μs`

For this dataset the RLE scheme improves predicate evauation by **88%** from `630μs` down to `70μs`. This performance improvement is realised primarily because the predicate can be evaluated on the compressed representation directly (the run-lengths).

### Row Materialisation

Finally I tested how quickly a column encoded with this scalar run-lengtrh encoding was able to materialise compared to the existing scheme, where float values were effectively held in a `Vec<T>`.

A simple way to look at this is with the equivalent of:

```
SELECT "container_file_descriptors" FROM "prometheus";
```

which for the RLE case forces the entire column to be decompressed and materialised.

|  | Before | RLE PR |
| --- | --- | -- |
| Execution Time | `901μs` | `966μs`

These results show that the RLE scheme is not significantly slower (about 7%) to materialise values than a simple `Vec<T>`. I haven't dug in too deeply into this yet but it's certainly not terribly worse given then space improvements.

## Next Steps

I have a lot of next steps! 

Here is a good first one... As mentioned, this RLE implementation uses `(u32, Option<T>)` to store run-lengths, which for `T = f64` means each run-length requires `24b`. One thing I want to do is to detect when a Float column contains only _natural numbers_ because then that means we can encode the column as integers. This is good for two reasons:

 (1) we can take advantage of the byte trimming functionality in the Read Buffer, which allows us to store a column of i64 as, e.g., u8 if all the values are in range;
 (2) more imporantly for RLE, if the values allow then we can make use of `NonZeroX` and friends. These will be very helpful because it will mean that `Option<T>` effectively becomes free and we can reduce the size of, e.g., ``(u32, Option<i64>)` from `24b` to `(u32, Option<NonZeroI64>)`, which is only `16b`.