@mkmik 2021-05-14
> The implementation is relatively straightforward, e.g., I have used (u32, Option<T>) for storing run-lengths. Naturally that can be improved upon, but this PR is following the 80/20 rule and trying to get the majority of the value added more quickly.

it's not clear to me where you actually use this tuple type, on a cursory look I only see it in the size estimation code:

```
let rle_size = rle_rows(arr) * size_of::<(u32, Option<f64>)>(); // size of a run length
```

Anyway, if we actually use that tuple, consider that:

`sizeof (u32, Option<f64>)` is 24 bytes

`sizeof (u32, bool, f64)` is only 16 bytes

perhaps there is some room for improvement (not necessarily now, this PR seems a pretty decent low hanging fruit anyway)