@e-dard 2021-05-14
I think I got almost all the comments. I also needed to add an uncompressed size estimator, which is here: https://github.com/influxdata/influxdb_iox/pull/1493/files#diff-e665da1718dd2b0069c3fdde70c3987b6b9bb60a2450c0ae61785ff20e78a6b4R91-R104

I will likely merge this soon.